services/analysis/requirements.txt
crewai
flask
python-dotenv

# If CrewAI uses OpenAI under the hood, you may also need:
# openai